+++ src/engine/incremental/env_state/fs.rs
@@ -1,4 +1,4 @@
-use super::super::is_in_checksums_dir;
+use crate::work_dir;
 use anyhow::{Context, Result};
 use rayon::prelude::*;
 use seahash::SeaHasher;
@@ -57,7 +57,7 @@ fn list_files(paths: &[PathBuf]) -> HashSet<PathBuf> {
                 Err(e) => log::debug!("Failed to walk dir {}: {}", path.display(), e),
                 Ok(entry) => {
                     let path = entry.path().to_path_buf();
-                    if path.is_file() && !is_in_checksums_dir(&path) {
+                    if path.is_file() && !work_dir::is_in_work_dir(&path) {
                         files.insert(path);
                     }
                 }
